462-machine-learning/hw3/requirements.txt
numpy==1.19.1
matplotlib==3.3.4
pandas==1.1.3
libsvm-official
